split of GameData
Had to reorganize my biggest class.
allowed to see more raid suggestions on (still hypothetical...) mac version
