+++ Shared/Summary/SummaryOfNotableRaids.swift
@@ -12,6 +12,8 @@ struct SummaryOfNotableRaids: View {
     @Namespace var animationsForSecondWidget
     @Namespace var animationsForThirdWidget
     @Namespace var animationsForFourthWidget
+    @Namespace var animationsForFifthWidget
+    @Namespace var animationsForSixthWidget
     
     let summarySize: summaryPreviewSize
     let character: CharacterInProfile
@@ -125,6 +127,10 @@ struct SummaryOfNotableRaids: View {
     
     func selectNamespace(for widgetNumber: Int) -> Namespace.ID {
         switch widgetNumber {
+        case 5:
+            return animationsForSixthWidget
+        case 4:
+            return animationsForFifthWidget
         case 3:
             return animationsForFourthWidget
         case 2:
